Refactor: MVC 패턴으로 클래스 변경
