+++ src/main/java/racingcar/Application.java
@@ -1,9 +1,11 @@
 package racingcar;
 
+import racingcar.controller.GameController;
+
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        Game game = new Game();
-        game.play();
+        GameController gameController = new GameController();
+        gameController.play();
     }
 }
