+++ src/test/java/racingcar/InputViewTest.java
@@ -1,7 +1,8 @@
-package racingcar.view;
+package racingcar;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+import racingcar.view.InputView;
 
 import java.util.ArrayList;
 import java.util.List;
